userinfo/requirements.txt
asgiref==3.2.3
Django==3.0
django-crispy-forms==1.7.2
djangorestframework==3.10.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
PyJWT==1.7.1
pytz==2019.1
sqlparse==0.3.0
